requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0

# Validación de datos
pydantic==2.5.0

# Utilidades
python-multipart==0.0.6

# Desarrollo (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2